+++ Classes/Domain/Repository/FileRepository.php
@@ -1,8 +1,11 @@
 <?php
+
 namespace Schmitzal\Tinyimg\Domain\Repository;
 
 use Schmitzal\Tinyimg\Domain\Model\FileStorage;
 use TYPO3\CMS\Extbase\Persistence\Repository;
+use TYPO3\CMS\Extbase\Persistence\QueryInterface;
+use TYPO3\CMS\Extbase\Persistence\QueryResultInterface;
 
 /**
  * Class FileRepository
@@ -13,7 +16,7 @@ class FileRepository extends Repository
     /**
      * Do not respect storage pid for domain records
      */
-    public function createQuery()
+    public function createQuery(): QueryInterface
     {
         $query = parent::createQuery();
         $query->getQuerySettings()->setRespectStoragePage(false);
@@ -24,19 +27,32 @@ class FileRepository extends Repository
     /**
      * @param FileStorage $storage
      * @param int $limit
+     * @param array $excludeFolders
      * @return array|\TYPO3\CMS\Extbase\Persistence\QueryResultInterface
      * @throws \TYPO3\CMS\Extbase\Persistence\Exception\InvalidQueryException
      */
-    public function findAllNonCompressedInStorageWithLimit(FileStorage $storage, $limit = 100)
+    public function findAllNonCompressedInStorageWithLimit(FileStorage $storage, $limit = 100, $excludeFolders = []): QueryResultInterface
     {
         $query = $this->createQuery();
+
+        $excludeFoldersConstraints = [];
+        foreach ($excludeFolders as $excludeFolder) {
+            $excludeFoldersConstraints[] = $query->logicalNot($query->like('identifier', $excludeFolder . '%'));
+        }
+
         $query->matching(
-            $query->logicalAnd([
-                $query->equals('storage', $storage),
-                $query->equals('compressed', false),
-                $query->equals('missing', false),
-                $query->in('extension', ['png', 'jpg', 'jpeg'])
-            ])
+            $query->logicalAnd(
+                array_merge(
+                    [
+                        $query->equals('storage', $storage),
+                        $query->equals('compressed', false),
+                        $query->equals('missing', false),
+                        $query->equals('compress_error', ''),
+                        $query->in('mime_type', ['image/png', 'image/jpeg'])
+                    ],
+                    $excludeFoldersConstraints
+                )
+            )
         );
         $query->setLimit($limit);