+++ Configuration/Extbase/Persistence/Classes.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    \Schmitzal\Tinyimg\Domain\Model\FileStorage::class => [
+        'tableName' => 'sys_file_storage'
+    ],
+    \Schmitzal\Tinyimg\Domain\Model\File::class => [
+        'tableName' => 'sys_file'
+    ]
+];