+++ Configuration/TypoScript/setup.ts
@@ -1,5 +1,0 @@
-module.tx_tinyimg.settings.debug = 0
-
-[applicationContext = Development]
-module.tx_tinyimg.settings.debug = 1
-[global]