+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### 1.5.0
+* **[FEATURE]** Add support for TYPO3 10.0 - Thanks to @achimfritz, @davidsteeb and @bmack from @b13 GmbH
+* **[CHANGE]** Drop support for TYPO3 7.6
+* **[CHANGE]** Rename TypoScript setup file
+* **[CHANGE]** Add compression errors and optimize flash messages
+
+### 1.4.0
+* **[FEATURE]** Add option to exclude specific folders from compression - Thanks to @achimfritz, @davidsteeb and @bmack from @b13 GmbH
+
 ### 1.3.0
 * **[FEATURE]** Add support for TYPO3 9.5
 * **[CHANGE]** Drop support of non composer installation